feat: cores iap
