+++ src/routes/webhooks/apple.ts
@@ -1,217 +1,39 @@
-import { readFile } from 'node:fs/promises';
-import { env } from 'node:process';
 import type { FastifyInstance, FastifyReply, FastifyRequest } from 'fastify';
 import { logger } from '../../logger';
 import {
   Environment,
   NotificationTypeV2,
   SignedDataVerifier,
-  Subtype,
   VerificationException,
   type JWSRenewalInfoDecodedPayload,
-  type ResponseBodyV2DecodedPayload,
 } from '@apple/app-store-server-library';
 import { concatTextToNewline, isTest, webhooks } from '../../common';
 import { isInSubnet } from 'is-in-subnet';
 import { isNullOrUndefined } from '../../common/object';
 import createOrGetConnection from '../../db';
-import {
-  UserSubscriptionStatus,
-  SubscriptionProvider,
-  User,
-  type UserSubscriptionFlags,
-} from '../../entity';
+import { SubscriptionProvider, User } from '../../entity';
 import { JsonContains } from 'typeorm';
-import { SubscriptionCycles } from '../../paddle';
-import { updateStoreKitUserSubscription } from '../../plusSubscription';
-import {
-  AnalyticsEventName,
-  sendAnalyticsEvent,
-} from '../../integrations/analytics';
 import type { Block, KnownBlock } from '@slack/web-api';
 import { remoteConfig } from '../../remoteConfig';
-import { convertCurrencyToUSD } from '../../integrations/openExchangeRates';
-
-const certificatesToLoad = isTest
-  ? ['__tests__/fixture/testCA.der']
-  : [
-      '/usr/local/share/ca-certificates/AppleIncRootCertificate.cer',
-      '/usr/local/share/ca-certificates/AppleRootCA-G2.cer',
-      '/usr/local/share/ca-certificates/AppleRootCA-G3.cer',
-    ];
-
-const loadAppleRootCAs = async (): Promise<Buffer[]> => {
-  const appleRootCAs: Buffer[] = await Promise.all(
-    certificatesToLoad.map(async (certPath) => await readFile(certPath)),
-  );
-
-  logger.debug(`Loaded ${appleRootCAs.length} Apple's Root CAs`);
-  return appleRootCAs;
-};
-
-const getVerifierEnvironment = (): Environment => {
-  switch (env.NODE_ENV) {
-    case 'production':
-      return Environment.PRODUCTION;
-    case 'development':
-      return Environment.SANDBOX;
-    case 'test':
-      return Environment.LOCAL_TESTING;
-    default:
-      throw new Error("Invalid 'NODE_ENV' value");
-  }
-};
-
-const bundleId = isTest ? 'dev.fylla' : env.APPLE_APP_BUNDLE_ID;
-const appAppleId = parseInt(env.APPLE_APP_APPLE_ID);
-const enableOnlineChecks = true;
-const environment = getVerifierEnvironment();
-
-const allowedIPs = [
-  '127.0.0.1/24',
-  '192.168.0.0/16',
-  '172.16.0.0/12',
-  '10.0.0.0/8',
-
-  // Production IPs. These are the IPs that Apple uses to send notifications.
-  // https://developer.apple.com/documentation/appstoreservernotifications/enabling-app-store-server-notifications#Configure-an-allow-list
-  '17.0.0.0/8',
-];
-
-interface AppleNotificationRequest {
-  signedPayload: string;
-}
-
-const productIdToCycle = {
-  annualSpecial: SubscriptionCycles.Yearly,
-  annual: SubscriptionCycles.Yearly,
-  monthly: SubscriptionCycles.Monthly,
-};
-
-const renewalInfoToSubscriptionFlags = (
-  data: JWSRenewalInfoDecodedPayload,
-): UserSubscriptionFlags => {
-  const cycle =
-    productIdToCycle[data.autoRenewProductId as keyof typeof productIdToCycle];
-
-  if (isNullOrUndefined(cycle)) {
-    logger.error(
-      { data, provider: SubscriptionProvider.AppleStoreKit },
-      'Invalid auto renew product ID',
-    );
-    throw new Error('Invalid auto renew product ID');
-  }
-
-  return {
-    cycle,
-    subscriptionId: data.originalTransactionId,
-    createdAt: new Date(data.recentSubscriptionStartDate!),
-    expiresAt: new Date(data.renewalDate!),
-    provider: SubscriptionProvider.AppleStoreKit,
-  };
-};
-
-const getSubscriptionStatus = (
-  notificationType: ResponseBodyV2DecodedPayload['notificationType'],
-  subtype?: ResponseBodyV2DecodedPayload['subtype'],
-): UserSubscriptionStatus => {
-  switch (notificationType) {
-    case NotificationTypeV2.SUBSCRIBED:
-    case NotificationTypeV2.DID_RENEW:
-    case NotificationTypeV2.DID_CHANGE_RENEWAL_PREF: // Upgrade/Downgrade
-    case NotificationTypeV2.REFUND_REVERSED:
-      return UserSubscriptionStatus.Active;
-    case NotificationTypeV2.DID_CHANGE_RENEWAL_STATUS: // Disable/Enable Auto-Renew
-      return subtype === Subtype.AUTO_RENEW_ENABLED
-        ? UserSubscriptionStatus.Active
-        : UserSubscriptionStatus.Cancelled;
-    case NotificationTypeV2.DID_FAIL_TO_RENEW:
-      // When user fails to renew and there is no grace period
-      if (isNullOrUndefined(subtype)) {
-        return UserSubscriptionStatus.Cancelled;
-      }
-    case NotificationTypeV2.GRACE_PERIOD_EXPIRED:
-      return UserSubscriptionStatus.Cancelled;
-    case NotificationTypeV2.REFUND:
-    case NotificationTypeV2.EXPIRED:
-    case NotificationTypeV2.REVOKE: // We don't support Family Sharing, but to be on the safe side
-      return UserSubscriptionStatus.Expired;
-    default:
-      logger.error(
-        {
-          notificationType,
-          subtype,
-          provider: SubscriptionProvider.AppleStoreKit,
-        },
-        'Unknown notification type',
-      );
-      throw new Error('Unknown notification type');
-  }
-};
-
-const getSubscriptionAnalyticsEvent = (
-  notificationType: ResponseBodyV2DecodedPayload['notificationType'],
-  subtype?: ResponseBodyV2DecodedPayload['subtype'],
-): AnalyticsEventName | null => {
-  switch (notificationType) {
-    case NotificationTypeV2.SUBSCRIBED:
-    case NotificationTypeV2.DID_RENEW:
-      return AnalyticsEventName.ReceivePayment;
-    case NotificationTypeV2.DID_CHANGE_RENEWAL_PREF:
-      return AnalyticsEventName.ChangeBillingCycle;
-    case NotificationTypeV2.DID_CHANGE_RENEWAL_STATUS: // Disable/Enable Auto-Renew
-      return subtype === Subtype.AUTO_RENEW_ENABLED
-        ? null
-        : AnalyticsEventName.CancelSubscription;
-    case NotificationTypeV2.DID_FAIL_TO_RENEW:
-      // When user fails to renew and there is no grace period
-      if (isNullOrUndefined(subtype)) {
-        return AnalyticsEventName.CancelSubscription;
-      }
-    default:
-      return null;
-  }
-};
-
-export const logAppleAnalyticsEvent = async (
-  data: JWSRenewalInfoDecodedPayload,
-  eventName: AnalyticsEventName,
-  user: User,
-  currencyInUSD: number,
-) => {
-  if (!data || isTest) {
-    return;
-  }
-
-  const cycle =
-    productIdToCycle[data?.autoRenewProductId as keyof typeof productIdToCycle];
-  const cost = data?.renewalPrice;
-
-  const extra = {
-    payment: SubscriptionProvider.AppleStoreKit,
-    cycle,
-    cost: currencyInUSD,
-    currency: 'USD',
-    localCost: cost ? cost / 1000 : undefined,
-    localCurrency: data?.currency,
-    payout: {
-      total: currencyInUSD * 100,
-      grandTotal: currencyInUSD * 100,
-      currencyCode: 'USD',
-    },
-  };
-
-  await sendAnalyticsEvent([
-    {
-      event_name: eventName,
-      event_timestamp: new Date(data?.signedDate || ''),
-      event_id: data.appTransactionId,
-      app_platform: 'api',
-      user_id: user.id,
-      extra: JSON.stringify(extra),
-    },
-  ]);
-};
+import {
+  getAppleTransactionType,
+  verifyAndDecodeAppleSignedData,
+} from '../../common/apple/utils';
+import {
+  allowedIPs,
+  appAppleId,
+  bundleId,
+  appleEnableOnlineChecks,
+  appleEnvironment,
+  type AppleNotificationRequest,
+  AppleTransactionType,
+} from '../../common/apple/types';
+import {
+  handleCoresPurchase,
+  isCorePurchaseApple,
+} from '../../common/apple/purchase';
+import { handleAppleSubscription } from '../../common/apple/subscription';
+import { loadAppleRootCAs } from '../../common/apple/utils';
 
 const handleNotifcationRequest = async (
   verifier: SignedDataVerifier,
@@ -250,30 +72,22 @@ const handleNotifcationRequest = async (
       return response.status(200).send({ received: true });
     }
 
-    // Check if the event is a subscription event
-    // NOTE: When adding support for purchasing cores, we must remove this check as it's not a subscription event
-    if (isNullOrUndefined(notification?.data?.signedRenewalInfo)) {
-      logger.info(
-        {
-          environment,
-          notification,
-          provider: SubscriptionProvider.AppleStoreKit,
-        },
-        "Missing 'signedRenewalInfo' in notification data",
-      );
+    const decodedInfo = await verifyAndDecodeAppleSignedData({
+      notification,
+      environment,
+      verifier,
+    });
+
+    if (!decodedInfo) {
       return response.status(400).send({ error: 'Invalid Payload' });
     }
-
-    const renewalInfo = await verifier.verifyAndDecodeRenewalInfo(
-      notification.data.signedRenewalInfo!,
-    );
 
     const con = await createOrGetConnection();
     const user = await con.getRepository(User).findOne({
       select: ['id', 'subscriptionFlags'],
       where: {
         subscriptionFlags: JsonContains({
-          appAccountToken: renewalInfo.appAccountToken,
+          appAccountToken: decodedInfo.appAccountToken,
         }),
       },
     });
@@ -307,54 +121,34 @@ const handleNotifcationRequest = async (
       return response.status(403).send({ error: 'Invalid Payload' });
     }
 
-    // Prevent double subscription
-    if (user.subscriptionFlags?.provider === SubscriptionProvider.Paddle) {
-      logger.error(
-        {
+    switch (getAppleTransactionType({ decodedInfo })) {
+      case AppleTransactionType.Consumable:
+        if (isCorePurchaseApple({ decodedInfo })) {
+          await handleCoresPurchase({
+            decodedInfo,
+            user,
+            environment,
+            notification,
+          });
+        } else {
+          throw new Error('Unsupported Apple Consumable transaction type');
+        }
+        break;
+      case AppleTransactionType.AutoRenewableSubscription:
+        await handleAppleSubscription({
+          decodedInfo,
           user,
           environment,
           notification,
-          provider: SubscriptionProvider.AppleStoreKit,
-        },
-        'User already has a Paddle subscription',
-      );
-      throw new Error('User already has a Paddle subscription');
-    }
-
-    const subscriptionStatus = getSubscriptionStatus(
-      notification.notificationType,
-      notification.subtype,
-    );
-
-    const eventName = getSubscriptionAnalyticsEvent(
-      notification.notificationType,
-      notification.subtype,
-    );
-
-    const subscriptionFlags = renewalInfoToSubscriptionFlags(renewalInfo);
-
-    await updateStoreKitUserSubscription({
-      userId: user.id,
-      status: subscriptionStatus,
-      data: subscriptionFlags,
-    });
-
-    const currencyInUSD = await convertCurrencyToUSD(
-      (renewalInfo.renewalPrice || 0) / 1000,
-      renewalInfo.currency || 'USD',
-    );
-
-    if (eventName) {
-      await logAppleAnalyticsEvent(renewalInfo, eventName, user, currencyInUSD);
-    }
-
-    if (notification.notificationType === NotificationTypeV2.SUBSCRIBED) {
-      await notifyNewStoreKitSubscription(renewalInfo, user, currencyInUSD);
+        });
+        break;
+      default:
+        throw new Error('Unsupported Apple transaction type');
     }
 
     logger.info(
       {
-        renewalInfo,
+        decodedInfo,
         user,
         environment,
         provider: SubscriptionProvider.AppleStoreKit,
@@ -509,13 +303,18 @@ export const apple = async (fastify: FastifyInstance): Promise<void> => {
     ) => {
       const verifier = new SignedDataVerifier(
         appleRootCAs,
-        enableOnlineChecks,
-        environment,
+        appleEnableOnlineChecks,
+        appleEnvironment,
         bundleId,
         appAppleId,
       );
 
-      await handleNotifcationRequest(verifier, request, response, environment);
+      await handleNotifcationRequest(
+        verifier,
+        request,
+        response,
+        appleEnvironment,
+      );
     },
   );
 
@@ -527,7 +326,7 @@ export const apple = async (fastify: FastifyInstance): Promise<void> => {
     ) => {
       const verifier = new SignedDataVerifier(
         appleRootCAs,
-        enableOnlineChecks,
+        appleEnableOnlineChecks,
         Environment.SANDBOX,
         bundleId,
         appAppleId,
